fix contacts page
